ppc64: Fix LXVL[L]/STXVL[L] memory block
EA is not calculated with "reg plus reg" for these instructions.
Instead, RA is the base register to access up to 16 bytes, according
to RB[0:7] value. Also, abuse the memory block to put a valid value in
RB.
--
Power ISA says that the results are "boundedly undefined" if the
contents of bits 8:63 of RB are not equal to zero. I'd expect the
hardware to ignore those bits, but we observed SIGSEGVs on hardware if
we use just reg($ra) and let any random value in RB.